"Added JSDoc comments and type annotations to various files, including controllers, models, and utility files. Refactored some code, including the admin and user routers, and updated the database connection and email transporter setup."
