+++ controllers/admins/courses.js
@@ -1,8 +1,13 @@
 import Course from "../../models/Course.js";
-import { unlink } from "fs";
 import { Op, Sequelize } from "sequelize";
 import Topic from "../../models/Topic.js";
 import { put, del } from "@vercel/blob";
+
+/**
+ * Admin create course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 
 export const createCourse = async (req, res) => {
   try {
@@ -42,6 +47,11 @@ export const createCourse = async (req, res) => {
   }
 };
 
+/**
+ * Admin search all course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const searchAllCourses = async (req, res) => {
   try {
     let { q: searchQuery } = req.query;
@@ -77,6 +87,11 @@ export const searchAllCourses = async (req, res) => {
   }
 };
 
+/**
+ * Admin get all course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const getAllCourse = async (req, res) => {
   try {
     let courses = await Course.findAll({
@@ -97,6 +112,11 @@ export const getAllCourse = async (req, res) => {
   }
 };
 
+/**
+ * Admin search all course by category controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const getAllCourseByCategory = async (req, res) => {
   try {
     let { category } = req.params;
@@ -122,45 +142,11 @@ export const getAllCourseByCategory = async (req, res) => {
   }
 };
 
-export const createTopic = async (req, res) => {
-  try {
-    let { courseId } = req.params;
-    let { title, note, description } = req.body;
-    let { buffer, mimetype } = req.file;
-
-    const { url } = await put(`/videos/video`, buffer, {
-      contentType: mimetype,
-      access: "public",
-    });
-
-    const topic = await Topic.create({
-      title,
-      note,
-      description,
-      courseId: courseId,
-      video: url,
-    });
-
-    await topic.save();
-
-    if (!topic) {
-      return res.status(401).json({
-        message: "Error creating topic",
-      });
-    }
-
-    res.status(200).json({
-      message: "Topic created successfully",
-    });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({
-      message:
-        "Error creating topic, This may be because you passed an invalid courseId",
-    });
-  }
-};
-
+/**
+ * Admin edit course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const editCourse = async (req, res) => {
   try {
     let path;
@@ -214,57 +200,11 @@ export const editCourse = async (req, res) => {
   }
 };
 
-export const editTopic = async (req, res) => {
-  try {
-    let path;
-    let { topicId } = req.params;
-    let { title, note, description } = req.body;
-
-    let topic = await Topic.findByPk(topicId);
-    if (!topic) {
-      return res.status(404).json({
-        message: "Topic not available or may be deleted",
-      });
-    }
-
-    if (req.file) {
-      await del(topic.video);
-
-      const { buffer, mimetype } = req.file;
-      path = await put(`/videos/video`, buffer, {
-        contentType: mimetype,
-        access: "public",
-      });
-
-      path = path.url;
-    } else {
-      path = topic.video;
-    }
-
-    topic.update({
-      title,
-      description,
-      note,
-      video: path,
-    });
-
-    if (topic.isNewRecord) {
-      return res.status(401).json({
-        message: "Error editing topic",
-      });
-    }
-
-    res.status(200).json({
-      message: "Topic edited successfully",
-    });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({
-      message: "Error editing topic",
-    });
-  }
-};
-
+/**
+ * Admin delete course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const deleteCourse = async (req, res) => {
   try {
     let { courseId } = req.params;
@@ -301,32 +241,11 @@ export const deleteCourse = async (req, res) => {
   }
 };
 
-export const deleteTopic = async (req, res) => {
-  try {
-    let { topicId } = req.params;
-
-    let topic = await Topic.findByPk(topicId);
-
-    if (!topic)
-      return res.status(404).json({
-        message: "Topic not found",
-      });
-
-    let { video } = topic;
-    await topic.destroy({ force: true });
-    await del(video);
-
-    res.status(200).json({
-      message: "Topic deleted successfully",
-    });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({
-      message: "Error deleting topic",
-    });
-  }
-};
-
+/**
+ * Admin publish course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const publishCourse = async (req, res) => {
   try {
     let { courseId } = req.params;
@@ -357,6 +276,11 @@ export const publishCourse = async (req, res) => {
   }
 };
 
+/**
+ * Admin unpublish course controller
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ */
 export const unpublishCourse = async (req, res) => {
   try {
     let { courseId } = req.params;
