+++ migrate.js
@@ -4,6 +4,35 @@ import Report from "./models/Report.js";
 import User from "./models/User.js";
 import { database } from "./util/util.js";
 
+// script.js
+const args = process.argv.slice(2); // Get all arguments after the first two
+
+const flags = {};
+args.forEach((arg) => {
+  const [key, value] = arg.split("=");
+  if (value) {
+    flags[key.replace("--", "")] = value;
+  } else {
+    flags[key.replace("--", "")] = true; // Flag without value
+  }
+});
+
+/**
+ * Migrate all models to the database.
+ * 
+ * If your models have associations create them in this function.
+ * Example:
+ * ```javascript 
+ *  // User-Report relation
+    User.hasMany(Report, {
+      onDelete: "SET NULL",
+      onUpdate: "NO ACTION",
+    });
+    Report.belongsTo(User);
+ * ```
+ * 
+ * To run function from CLI run `node migrate --development`
+ */
 const migrate = async () => {
   // User-Report relation
   User.hasMany(Report, {
@@ -31,19 +60,6 @@ const migrate = async () => {
 
   console.log("\n✅ All model synced");
 };
-
-// script.js
-const args = process.argv.slice(2); // Get all arguments after the first two
-
-const flags = {};
-args.forEach((arg) => {
-  const [key, value] = arg.split("=");
-  if (value) {
-    flags[key.replace("--", "")] = value;
-  } else {
-    flags[key.replace("--", "")] = true; // Flag without value
-  }
-});
 
 if (flags.development) await migrate();
